Fixing unit tests for new expostal API
